refactor: separate Influencer Card component from Homepage
