+++ src/pages/user/home/HomePage.tsx
@@ -14,7 +14,9 @@ import {
   itemMockData,
   recommendMockData,
 } from '@/pages/user/home/HomeMockData';
-import cn from '@/utils/cn';
+import InfluencerCard, {
+  InfluencerCardType,
+} from '@/components/user/home/InfluencerCard';
 
 interface TopBannerItem {
   image: string;
@@ -44,14 +46,7 @@ export const topBannerMockData: TopBannerItem[] = [
   },
 ];
 
-interface Influencer {
-  id: number;
-  nickname: string;
-  username: string;
-  profileImage: string;
-}
-
-export const influencerMockData: Influencer[] = [
+export const influencerMockData: InfluencerCardType[] = [
   {
     id: 1,
     nickname: '혜선',
@@ -119,7 +114,7 @@ const HomePage = () => {
   const [selectedInfluencer, setSelectedInfluencer] = useState<number>(1);
 
   return (
-    <section className="top-banner-swiper-section bg-grey01 scrollbar-hide relative flex w-full flex-1 flex-col overflow-x-hidden overflow-y-auto">
+    <section className="top-banner-swiper-section bg-grey01 flex w-full flex-1 flex-col">
       <PageHeader
         leftIcons={[<InfluyIcon role="button" aria-label="뒤로 가기" />]}
         rightIcons={[
@@ -129,96 +124,62 @@ const HomePage = () => {
             <div className="absolute top-0.5 right-[.2188rem] h-1.5 w-1.5 rounded-full bg-[#F43232]" />
           </button>,
         ]}
-        additionalStyles="bg-white border-0"
-      >
-        <div className="h-[1.6875rem]" />
-      </PageHeader>
-      <TopBannerSwiper data={topBannerMockData} />
-      <section className="flex w-full flex-col gap-4 pt-7 pb-3">
-        <h1 className="subhead-b px-5 text-black">내 취향의 인플루언서</h1>
-        <div className="scrollbar-hide flex gap-6 overflow-x-scroll px-5">
-          {influencerMockData.map((influencer) => (
-            <InfluencerCard
-              key={influencer.id}
-              influencer={influencer}
-              selectedInfluencer={selectedInfluencer}
-              setSelectedInfluencer={setSelectedInfluencer}
-            />
-          ))}
-        </div>
-      </section>
-      {/* 셀러가 픽한 상품 */}
-      <section className="bg-grey02 mb-5 flex w-full flex-col gap-[.875rem] pt-4 pb-5">
-        {/* 제목 */}
-        <div className="flex items-center justify-between px-5">
-          <h1 className="subhead-b text-black">
-            {
-              influencerMockData.find(
-                (influencer) => influencer.id === selectedInfluencer
-              )?.nickname
-            }
-            님이 <span className="text-main">Pick</span>한 상품
-          </h1>
-          <MoreButton onClickMore={() => {}} />
-        </div>
-
-        {/* 사진 */}
-        <div className="flex w-full gap-0.5 px-5">
-          {pickMockData.map((item) => (
-            <div className="aspect-square flex-1/3">
-              <img
-                src={item.image}
-                className="h-full w-full rounded-[.1273rem] object-cover"
-              />
-            </div>
-          ))}
-        </div>
-      </section>
-      <HomeCommonSection
-        expiringItem={itemMockData}
-        trendingItem={itemMockData}
-        recommendedItem={recommendMockData}
-        categoryList={dummyCategory}
-        selectedCategory={selectedCategory}
-        setSelectedCategory={setSelectedCategory}
+        additionalStyles="border-0 h-11"
       />
+      <section className="scrollbar-hide flex w-full flex-1 flex-col overflow-x-hidden overflow-y-auto">
+        <TopBannerSwiper data={topBannerMockData} />
+        <section className="flex w-full flex-col gap-4 pt-7 pb-3">
+          <h1 className="subhead-b px-5 text-black">내 취향의 인플루언서</h1>
+          <ul className="scrollbar-hide flex gap-6 overflow-x-scroll px-5">
+            {influencerMockData.map((influencer) => (
+              <InfluencerCard
+                key={influencer.id}
+                influencer={influencer}
+                selectedInfluencer={selectedInfluencer}
+                setSelectedInfluencer={setSelectedInfluencer}
+              />
+            ))}
+          </ul>
+        </section>
+        {/* 셀러가 픽한 상품 */}
+        <section className="bg-grey02 mb-5 flex w-full flex-col gap-[.875rem] pt-4 pb-5">
+          {/* 제목 */}
+          <div className="flex items-center justify-between px-5">
+            <h1 className="subhead-b text-black">
+              {
+                influencerMockData.find(
+                  (influencer) => influencer.id === selectedInfluencer
+                )?.nickname
+              }
+              님이 <span className="text-main">Pick</span>한 상품
+            </h1>
+            <MoreButton onClickMore={() => {}} />
+          </div>
+
+          {/* 사진 */}
+          <div className="flex w-full gap-0.5 px-5">
+            {pickMockData.map((item) => (
+              <div className="aspect-square flex-1/3">
+                <img
+                  src={item.image}
+                  className="h-full w-full rounded-[.1273rem] object-cover"
+                />
+              </div>
+            ))}
+          </div>
+        </section>
+        <HomeCommonSection
+          expiringItem={itemMockData}
+          trendingItem={itemMockData}
+          recommendedItem={recommendMockData}
+          categoryList={dummyCategory}
+          selectedCategory={selectedCategory}
+          setSelectedCategory={setSelectedCategory}
+        />
+      </section>
       <BottomNavBar />
     </section>
   );
 };
 
 export default HomePage;
-
-const InfluencerCard = ({
-  influencer,
-  selectedInfluencer,
-  setSelectedInfluencer,
-}: {
-  influencer: Influencer;
-  selectedInfluencer: number;
-  setSelectedInfluencer: (id: number) => void;
-}) => {
-  return (
-    <article
-      className="flex flex-col items-center gap-2"
-      onClick={() => setSelectedInfluencer(influencer.id)}
-    >
-      <img
-        src={influencer.profileImage}
-        alt="프로필 사진"
-        className={cn(
-          'aspect-square h-[3.75rem] rounded-full object-cover',
-          selectedInfluencer == influencer.id && 'border-main border-[.0938rem]'
-        )}
-      />
-      <div className="flex w-[3.75rem] flex-col items-center self-stretch text-[.8086rem] leading-[150%] tracking-[-0.0008rem]">
-        <p className="text-grey10 line-clamp-1 font-semibold">
-          {influencer.nickname}
-        </p>
-        <p className="text-grey09 line-clamp-1 font-normal">
-          {influencer.username}
-        </p>
-      </div>
-    </article>
-  );
-};
